% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_eirrc.R
\name{fit_eirrc}
\alias{fit_eirrc}
\title{fit_eirrc}
\usage{
fit_eirrc(
  obs,
  time,
  priors_only,
  n_samples,
  n_chains,
  seed,
  gamma_sd = 0.2,
  gamma_mean = log(1/4),
  nu_sd = 0.2,
  nu_mean = log(1/7),
  eta_sd = 0.2,
  eta_mean = log(1/18),
  rho_gene_sd = 1,
  rho_gene_mean = 0,
  tau_sd = 1,
  tau_mean = 0,
  I_init_sd = 0.05,
  I_init_mean = 489,
  R1_init_sd = 0.05,
  R1_init_mean = 2075,
  E_init_sd = 0.05,
  E_init_mean = 225,
  lambda_mean = 5.685528,
  lambda_sd = 2.178852,
  df_shape = 2,
  df_scale = 10,
  sigma_rt_sd = 0.2,
  sigma_rt_mean = log(0.1),
  rt_init_sd = 0.1,
  rt_init_mean = log(0.88)
)
}
\arguments{
\item{obs}{real valued vector of observed RNA concentrations on log scal}

\item{time}{real valued vector of observation times, corresponds to obs vector}

\item{priors_only}{boolean if sample from the prior distribution only}

\item{n_samples}{integer number of desired posterior samples,  model will run 2*n_samples total samples and discard half as burn-in}

\item{n_chains}{number of chains to run, note you'll need to specify number of threads available separately}

\item{seed}{set the seed}

\item{gamma_sd}{real valued sd for gamma prior on log scale}

\item{gamma_mean}{real valued mean for gamma prior on log scale}

\item{nu_sd}{real valued sd for nu prior on log scale}

\item{nu_mean}{real valued mean for nu prior}

\item{eta_sd}{real valued sd for eta prior on log scale}

\item{eta_mean}{real valued mean for eta prior on log scale}

\item{rho_gene_sd}{real valued sd for rho prior on log scale}

\item{rho_gene_mean}{real valued mean for rho prior on log scale}

\item{tau_sd}{real valued sd for tau prior on log scale}

\item{tau_mean}{real valued mean for tau prior on log scale}

\item{I_init_sd}{real valued sd for initial I}

\item{I_init_mean}{real valued mean for intial I}

\item{R1_init_sd}{real valued sd for initial R1}

\item{R1_init_mean}{real valued mean for initial R1}

\item{E_init_sd}{real valued sd for initial E}

\item{E_init_mean}{real valued mean for initial E}

\item{lambda_mean}{real valued mean for lambda on logit scale}

\item{lambda_sd}{real valued sd for lambda on logit scale}

\item{df_shape}{real valued shape for df prior}

\item{df_scale}{real valued scale for df prior}

\item{sigma_rt_sd}{real valued sd for sigma rt on log scale}

\item{sigma_rt_mean}{real valued sd for sigma rt on log scale}

\item{rt_init_sd}{real valued sd for initial rt on log scale}

\item{rt_init_mean}{real valued mean for initial rt on log scale}
}
\description{
R wrapper for julia function fitting eirrc model to data. Priors default to priors used for scenario 1.
}
