% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_seirr.R
\name{fit_seirr}
\alias{fit_seirr}
\title{fit_seirr}
\usage{
fit_seirr(
  data,
  obstimes,
  param_change_times,
  extra_ode_precision,
  priors_only,
  n_samples = 250L,
  n_chains = 4L,
  seed = 1L,
  fit_abs_tol = 1e-09,
  fit_rel_tol = 1e-06,
  opt_abs_tol = 1e-11,
  opt_rel_tol = 1e-08,
  popsize = 100000L,
  active_pop = 92271L,
  gamma_sd = 0.2,
  gamma_mean = log(1/4),
  nu_sd = 0.2,
  nu_mean = log(1/7),
  eta_sd = 0.2,
  eta_mean = log(1/18),
  rho_gene_sd = 1,
  rho_gene_mean = 0,
  tau_sd = 1,
  tau_mean = 0,
  sigma_R0_sd = 0.2,
  sigma_R0_mean = log(0.1),
  S_SEIR1_sd = 0.05,
  S_SEIR1_mean = 3.468354,
  I_EIR1_sd = 0.05,
  I_EIR1_mean = -1.548302,
  R1_ER1_sd = 0.05,
  R1_ER1_mean = 2.221616,
  r0_init_sd = 0.2,
  r0_init_mean = log(0.88),
  lambda_mean = 5.685528,
  lambda_sd = 2.178852,
  df_shape = 2,
  df_scale = 10
)
}
\arguments{
\item{data}{vector: log RNA concentrations}

\item{obstimes}{vector: times RNA concentrations are observed}

\item{param_change_times}{vector: times reproduction number is allowed to change}

\item{extra_ode_precision}{boolean: if TRUE, uses custom ode precisions, otherwise uses DifferentialEquations.jl default values}

\item{priors_only}{boolean: if true, function produces draws from the joint prior}

\item{n_samples}{integer: number of posterior samples AFTER burn-in, total will be twice n_samples}

\item{n_chains}{integer: number of chains}

\item{seed}{integer: random seed}

\item{fit_abs_tol}{float64: if `extra_ode_precision` TRUE, absolute tolerance for model fitting}

\item{fit_rel_tol}{float64: if `extra_ode_precision` TRUE, relative tolerance for model fitting}

\item{opt_abs_tol}{float64: if `extra_ode_precision` TRUE, absolute tolerance for choosing mcmc initial values}

\item{opt_rel_tol}{float64: if `extra_ode_precision` true, relative tolerance for choosing mcmc initial values}

\item{popsize}{integer: population size}

\item{active_pop}{intger: population size - initial size of R compartment}

\item{gamma_sd}{float64: standard deviation for normal prior of log gamma}

\item{gamma_mean}{float64: mean for normal prior of log gamma}

\item{nu_sd}{float64: standard deviation for normal prior of log nu}

\item{nu_mean}{float64: mean for normal prior of log nu}

\item{eta_sd}{float64: standard deviation for normal prior of log eta}

\item{eta_mean}{float64: mean for normal prior of log eta}

\item{rho_gene_sd}{float64: standard devation for normal prior of log rho}

\item{rho_gene_mean}{float64: mean for normal prior of log rho}

\item{tau_sd}{float64: standard deviation for normal prior of log tau}

\item{tau_mean}{float64: mean for normal prior of log tau}

\item{sigma_R0_sd}{flaot64: standard deviation for normal prior of log sigma R0}

\item{sigma_R0_mean}{float64: mean for normal prior of log sigma R0}

\item{S_SEIR1_sd}{float64: standard deviation for normal prior on logit fraction of `active_pop` initially in S}

\item{S_SEIR1_mean}{float64: mean for normal prior on logit fraction of `active_pop` initially in S}

\item{I_EIR1_sd}{float64: standard deviation for normal prior on logit fraction of initial E,I and R1 compartments in the I compartment}

\item{I_EIR1_mean}{float64: mean for normal prior on logit fraction of initial E,I and R1 compartments in the I compartment}

\item{R1_ER1_sd}{float64: standard deviation for normal prior on logit fraction of initial E and R1 compartments in the R1 compartment}

\item{R1_ER1_mean}{float64: mean for normal prior on logit fraction of initial E and R1 compartments in the R1 compartment}

\item{r0_init_sd}{float64: standard deviation for normal prior of log R0}

\item{r0_init_mean}{float64: mean for normal prior of log R0}

\item{lambda_mean}{float64: mean for normal prior of logit lambda}

\item{lambda_sd}{float64: standard deviation for normal prior of logit lambda}

\item{df_shape}{float64: shape parameter for gamma prior of df}

\item{df_scale}{float64: scale parameter for gamma prior of df}
}
\value{
Turing chains object
}
\description{
fit_seirr
}
