% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eir.R
\name{generate_eir}
\alias{generate_eir}
\title{generate_eir}
\usage{
generate_eir(
  samples,
  data_cases,
  obstimes,
  param_change_times,
  seed = 1,
  gamma_sd = 0.2,
  gamma_mean = log(1/4),
  nu_sd = 0.2,
  nu_mean = log(1/7),
  rho_case_sd = 1,
  rho_case_mean = 0,
  phi_sd = 0.2,
  phi_mean = log(50),
  I_init_sd = 0.05,
  I_init_mean = 489,
  E_init_sd = 0.05,
  E_init_mean = 225,
  sigma_rt_sd = 0.2,
  sigma_rt_mean = log(0.1),
  rt_init_sd = 0.1,
  rt_init_mean = log(0.88)
)
}
\arguments{
\item{samples}{output from `fit_eir`}

\item{data_cases}{vector: observed cases}

\item{obstimes}{vector: times cases are observed}

\item{param_change_times}{vector: times reproduction number is allowed to change}

\item{seed}{integer: random seed}

\item{gamma_sd}{float64: standard deviation for normal prior of log gamma}

\item{gamma_mean}{float64: mean for normal prior of log gamma}

\item{nu_sd}{float64: standard deviation for normal prior of log nu}

\item{nu_mean}{float64: mean for normal prior of log nu}

\item{rho_case_sd}{float64: standard devation for normal prior of log rho}

\item{rho_case_mean}{float64: mean for normal prior of log rho}

\item{phi_sd}{float64: standard deviation for normal prior of log phi}

\item{phi_mean}{float64: mean for normal prior of log phi}

\item{I_init_sd}{float64: standard deviation for normal prior of I_init}

\item{I_init_mean}{float64: mean for normal prior of I_init}

\item{E_init_sd}{float64: standard deviation for normal prior of E_init}

\item{E_init_mean}{float64: mean for normal prior of E_init}

\item{sigma_rt_sd}{float64: standard deviation for normal prior on log sigma rt}

\item{sigma_rt_mean}{float64: mean for normal prior on log sigma rt}

\item{rt_init_sd}{float64: standard deviation for normal prior on log rt_init}

\item{rt_init_mean}{float64: mean for normal prior on log rt_init}
}
\value{
List of three dataframes, posterior predictive, scaled posterior, and un-scaled posterior
}
\description{
#' Call `generate_pp_and_gq_eir` in julia to transform output of `fit_eir` to useable dataframes
}
\details{
Default priors are for scenario 1, and assume the model is being fit to a daily time scale
}
