% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eirrc.R
\name{generate_eirrc}
\alias{generate_eirrc}
\title{generate_eirrc}
\usage{
generate_eirrc(
  samples,
  obs,
  time,
  seed,
  gamma_sd = 0.2,
  gamma_mean = log(1/4),
  nu_sd = 0.2,
  nu_mean = log(1/7),
  eta_sd = 0.2,
  eta_mean = log(1/18),
  rho_gene_sd = 1,
  rho_gene_mean = 0,
  tau_sd = 1,
  tau_mean = 0,
  I_init_sd = 0.05,
  I_init_mean = 489,
  R1_init_sd = 0.05,
  R1_init_mean = 2075,
  E_init_sd = 0.05,
  E_init_mean = 225,
  lambda_mean = 5.685528,
  lambda_sd = 2.178852,
  df_shape = 2,
  df_scale = 10,
  sigma_rt_sd = 0.2,
  sigma_rt_mean = log(0.1),
  rt_init_sd = 0.1,
  rt_init_mean = log(0.88)
)
}
\arguments{
\item{samples}{a Turing chains object generated from fitting a Turing model}

\item{rt_init_mean}{}
}
\description{
Generate correctly scaled posterior samples, predictive distributions, and dataframe of un-scaled samples from Turing object.
The model must be reconstructed, so many parameters are re-used from fit_eirrc. If there is no description, check fit_eirrc for parameter explanation.
}
