% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_seir.R
\name{generate_seir}
\alias{generate_seir}
\title{generate_seir}
\usage{
generate_seir(
  samples,
  data_cases,
  obstimes,
  param_change_times,
  extra_ode_precision,
  seed = 1,
  fit_abs_tol = 1e-09,
  fit_rel_tol = 1e-06,
  popsize = 100000L,
  active_pop = 90196L,
  gamma_sd = 0.2,
  gamma_mean = log(7/4),
  nu_sd = 0.2,
  nu_mean = log(7/7),
  rho_case_sd = 0.4,
  rho_case_mean = -1.386294,
  phi_sd = 0.2,
  phi_mean = log(50),
  sigma_R0_sd = 0.2,
  sigma_R0_mean = log(0.1),
  S_SEI_sd = 0.05,
  S_SEI_mean = 4.83091,
  I_EI_sd = 0.05,
  I_EI_mean = 0.7762621,
  r0_init_sd = 0.1,
  r0_init_mean = log(0.88)
)
}
\arguments{
\item{samples}{output from `fit_seir`}

\item{data_cases}{vector: observed cases}

\item{obstimes}{vector: times cases are observed}

\item{param_change_times}{vector: times reproduction number is allowed to change}

\item{extra_ode_precision}{boolean: if TRUE, uses custom ode precisions, otherwise uses DifferentialEquations.jl default values}

\item{seed}{integer: random seed}

\item{fit_abs_tol}{float64: if `extra_ode_precision` TRUE, absolute tolerance for model fitting}

\item{fit_rel_tol}{float64: if `extra_ode_precision` TRUE, relative tolerance for model fitting}

\item{popsize}{integer: population size}

\item{active_pop}{intger: population size - initial size of R compartment}

\item{gamma_sd}{float64: standard deviation for normal prior of log gamma}

\item{gamma_mean}{float64: mean for normal prior of log gamma}

\item{nu_sd}{float64: standard deviation for normal prior of log nu}

\item{nu_mean}{float64: mean for normal prior of log nu}

\item{rho_case_sd}{float64: standard devation for normal prior of log rho}

\item{rho_case_mean}{float64: mean for normal prior of log rho}

\item{phi_sd}{float64: standard deviation for normal prior of log phi}

\item{phi_mean}{float64: mean for normal prior of log phi}

\item{sigma_R0_sd}{flaot64: standard deviation for normal prior of log sigma R0}

\item{sigma_R0_mean}{float64: mean for normal prior of log sigma R0}

\item{S_SEI_sd}{float64: standard deviation for normal prior of logit fraction of active pop initially in S}

\item{S_SEI_mean}{flaot64: mean for normal prior of logit fraction of active pop initially in S}

\item{I_EI_sd}{float64: standard deviation for normal prior of logit fraction of the E and I compartments initially in I}

\item{I_EI_mean}{float64: mean for normal prior of logit fraction of the E and I compartments initially in I}

\item{r0_init_sd}{float64: standard deviation for normal prior of log R0}

\item{r0_init_mean}{float64: mean for normal prior of log R0}
}
\value{
List of three dataframes, posterior predictive, scaled posterior, and un-scaled posterior
}
\description{
Call `generate_pp_and_gq_seir` in Julia to transform output from `fit_seir` to useable dataframes
}
\details{
Default priors are for scenario 1 and assume model is fit to a weekly time scale.
}
