% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fixed_samples.R
\name{make_fixed_samples}
\alias{make_fixed_samples}
\title{make_fixed_samples}
\usage{
make_fixed_samples(gq_samples)
}
\arguments{
\item{gq_samples}{dataframe of samples from Turing}
}
\description{
Create data frame of non-timevarying samples from Turing
}
