% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_timevarying_quantiles.R
\name{make_timevarying_quantiles}
\alias{make_timevarying_quantiles}
\title{make_timevarying_quantiles}
\usage{
make_timevarying_quantiles(gq_samples)
}
\arguments{
\item{gq_samples}{}
}
\value{
a dataframe
}
\description{
Make dataframe of 50,80 and 95 percent quantiles of time-varying parameters from Turing model.
The point estimate is the median.
}
