% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{setup}
\usage{
setup(install_julia = FALSE)
}
\arguments{
\item{install_julia}{if TRUE, install julia using `JuliaCall::install_julia`}
}
\value{
NA
}
\description{
Use`JuliaCall` to install julia if needed, and install `testpackage.jl` if needed.
}
